+++ src/main/java/subway/application/StationService.java
@@ -5,8 +5,9 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import subway.dao.StationDao;
 import subway.domain.Station;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.dto.request.StationRequest;
+import subway.dto.response.StationResponse;
+import subway.exception.NotFoundException;
 
 @Service
 public class StationService {
