+++ src/main/java/subway/dto/response/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 import java.util.List;
