+++ src/main/java/subway/entity/StationEntity.java
@@ -1,0 +1,4 @@
+package subway.entity;
+
+public class StationEntity {
+}
