+++ src/main/java/subway/ui/GlobalExceptionHandler.java
@@ -5,8 +5,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import subway.application.NotFoundException;
-import subway.dto.ExceptionResponse;
+import subway.exception.NotFoundException;
+import subway.dto.response.ExceptionResponse;
 import subway.exception.InvalidSectionException;
 import subway.exception.LineDuplicatedException;
