+++ src/main/java/subway/ui/SectionController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.SectionService;
-import subway.dto.SectionRequest;
+import subway.dto.request.SectionRequest;
 
 @RestController
 @RequestMapping("/sections")
